+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: remove duplicate Related Show box in Post Quick Edit
 * Fixed: multiple attributes for automatic pages shortcodes
 * Fixed: undefined warning in current playlist when no current show
+* Fixed: undefined warning for newline in schedule templates
 * Fixed: undefined warning for debugshifts
 
 ### 2.4.0.5